requirements.txt
# Project dependencies
requests
numpy
opencv-python
rasterio
scipy
matplotlib
python-dotenv
sentinelsat
geojson
pyproj
